Refactored login so that a CSRF token is used, also added the ability for a users session to be stored if they tick remember me
