+++ main/static/assets/css/styles.css
@@ -501,7 +501,7 @@ body {
 }
 
 /* Login pop up */
-input[type=text], input[type=password] {
+.inputField, input[type=text], input[type=password] {
   width: 100%;
   padding: 12px 20px;
   margin: 8px 0;
