+++ main/static/assets/css/styles.scss
@@ -530,7 +530,7 @@ body {
 /* Login pop up */
 
 // Full-width input fields
-input[type=text], input[type=password] {
+.inputField, input[type=text], input[type=password] {
     width: 100%;
     padding: 12px 20px;
     margin: 8px 0;
